Teszt fázis

+++ src/hu/ninedevs/pn/PlateNumbers.java
@@ -108,7 +108,14 @@ public class PlateNumbers {
     }
     static void appendPlateNumber(PlateNumber[] pn) throws IOException {
         File f = new File("rendszam.txt");
-        FileWriter fw = new FileWriter(f,true);
+        FileWriter fw;
+        if (f.exists()){
+            fw = new FileWriter(f,true);
+        }
+        else {
+            fw = new FileWriter(f,false);
+        }
+
         PrintWriter pw = new PrintWriter(fw);
         Issuer b = new Issuer(pn);
         for (int i = 0; i < 5; i++) {
